Bislice from two weeks ago
To walk the RGB (or cieXYZ) space, I want a depth walking algorithm and
this is my current stab at it. When I have it working, I'll move it to
java.

This and maybe a cheap Trie for LangLexer are next. I'm seeing that a
hashset is probably more appropriate than a Trie though. Oh well.