Next phase: pretend lexing
This version emits mangled html that only colors the code blocks. I'm
pleased but sleep deprived, which is why it mangles the code blocks. The
sine color space is meh. It moves pretty quickly around. SyntaxPDautoma
will handle actual lexing. LangLexer just selects the language for sPDa I
think? LitmusFrame reads a file the java7 way, but doesn't turn it into
transmix yet. That'll be a small change. Definitely not the way PD will,
so it's okay to have.

Bislice is now essential. sPDa needs to ignore operators, which will mean
a little more realism inside, but not yet stack use. The emitted styles
are huge. Maybe it's worth doing a pass and creating a page style sheet of
colors? Or have that as a config setting? Litmus badly needs the math of
splitting html blocks fixed. That would be enough for another 'milestone'
like this commit.