Bislicing ready
Bislice now has a depth first search algorithm, which I'll put into
colorspace. Part of my problem was python's default for [].pop() is to pop
the last when it has no argument. Well, I'm lucky it can pop from
anywhere.

Next is clearly to put this in colorspace, but I don't doubt I'll find
distractions in the other classes.